# backend/RestApiSwagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger PSWTM - OpenAPI 3.0",
    "description": "REST API Methods description for our Project",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:8800/api/v1"
    }
  ],
...
  "paths": {
...
    "/auth/loginRFID": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "POST Login User/Admin in Terminal",
        "description": "Login User/Admin in Terminal.",
        "operationId": "loginRFID",
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rfid": {
                    "type": "string",
                    "example": "521"
                  }
                },
                "xml": {
                  "name": "user"
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1
                    },
                    "credits": {
                      "type": "number",
                      "format": "float",
                      "example": "0"
                    },
                    "username": {
                      "type": "string",
                      "example": "John.James"
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJpYXQiOjE2NzM1NDg5NTgsImV4cCI6MTY3MzYzNTM1OH0.O_oxuGw4sSjS_J8pKfVIToQ3R48qIdZtIm07nmBjFMw"
                    },
                    "rfid": {
                      "type": "string",
                      "example": "521"
                    },
                    "is_admin": {
                      "type": "integer",
                      "description": "User Admin (1 yes, 0 no)",
                      "format": "int32",
                      "example": "false"
                    }
                  },
                  "xml": {
                    "name": "user"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "POST an order for a product",
        "description": "Place a new order in the Terminal",
        "operationId": "placeOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "integer",
                    "example": 10
                  },
                  "quantity": {
                    "type": "integer",
                    "example": 7
                  },
                  "product_name": {
                    "type": "string",
                    "example": "coffee"
                  },
                  "product_costs": {
                    "type": "number",
                    "format": "float",
                    "example": 2.5
                  }
                },
                "xml": {
                  "name": "order"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order_id": {
                      "type": "integer",
                      "example": 5
                    },
                    "user_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "product_id": {
                      "type": "integer",
                      "example": 10
                    },
                    "quantity": {
                      "type": "integer",
                      "example": 7
                    },
                    "product_name": {
                      "type": "string",
                      "example": "coffee"
                    },
                    "product_costs": {
                      "type": "integer",
                      "format": "float",
                      "example": 2.6
                    },
                    "date": {
                      "type": "string",
                      "description": "date in echter datenbank",
                      "format": "date-time",
                      "example": "15.11.2022"
                    }
                  },
                  "xml": {
                    "name": "order"
                  }
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "GET product history by date",
        "description": "Returns a map of status codes to quantities",
        "operationId": "getOrderByDate",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The user_id that needs to be fetched. Use 1 for testing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "The month that needs to be fetched. Use Oktober for testing.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/credits": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "GET credit history by date",
        "description": "Returns a map of status codes to quantities",
        "operationId": "getCreditByDate",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The user_id that needs to be fetched. Use 1 for testing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "The month that needs to be fetched. Use Oktober for testing.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}